fix: Move vars into shared file
