+++ apps/complete_registration/src/App.tsx
@@ -6,7 +6,6 @@ import { reverseMap } from "./utils/mapper"
 import { useDispatch } from "react-redux"
 import { setErrors, setField } from "./store/form/form_slice"
 import {
-    PACKAGE_KEY,
     ProductMeta,
     loadProducts,
     pickProduct
@@ -18,8 +17,7 @@ import {
     setCompanyRegistrationStatus,
     setUser
 } from "./store/company/company_slice"
-
-export const HOST = ""
+import { HOST, PACKAGE_KEY } from "./shared/vars"
 
 export function App() {
     const initialized = useRef(false)
@@ -34,11 +32,7 @@ export function App() {
             const data = await raw.json()
             dispatch(loadProducts(data))
         })
-        fetch(`${HOST}/api/registration/`, {
-            headers: new Headers({
-                "ngrok-skip-browser-warning": "69420"
-            })
-        }).then(async raw => {
+        fetch(`${HOST}/api/registration/`, {}).then(async raw => {
             const data = await raw.json()
 
             if (data.error != null) {
@@ -51,8 +45,8 @@ export function App() {
             dispatch(
                 setCompanyRegistrationStatus(data.type as RegistrationStatus)
             )
-            dispatch(setCompanyName(data.company.name))
-            dispatch(setUser(data.contact))
+            if (data.company?.name) dispatch(setCompanyName(data.company.name))
+            if (data.company?.contact) dispatch(setUser(data.contact))
 
             for (const current of awaitingMappings) {
                 dispatch(
@@ -65,7 +59,7 @@ export function App() {
 
             // Apply orders
             const orders = data.orders as ProductMeta[]
-            for (const productMeta of orders) {
+            for (const productMeta of orders ?? []) {
                 dispatch(
                     pickProduct({
                         id: productMeta.product.id,
