+++ apps/complete_registration/src/shared/vars.ts
@@ -1,0 +1,4 @@
+export const HOST = import.meta.env.DEV ? "http://localhost:3000" : ""
+export const PACKAGE_KEY = "Package"
+export const EVENTS_REGISTRATION_SECTION_KEY = "Events"
+export const EXTRAS_REGISTRATION_SECTION_KEY = "Extras"
